wip: Substantially flesh out handler abstraction
I ended up deleting the marker traits that distinguish between method
and notification as they were complicating things unnecessarily.  The
notification and method handler pathways are already separate, ensuring
the compiler can optimize the notification path with the knowledge that
the result is not used, so the only value the marker traits provided was
to encode the semantic difference in the type system.  That seems not
very worthwhile and possibly even undesirable, since conceivably a user
could have a method that has no return value and nonetheless think of it
as a method and not a notification handler.