wip: More progress on fleshing out JSON design
I've settled on an easy solution for transport: just don't solve that
problem at this layer.  The MCP impl is going to have some pretty
complex protocol-specific impls anyway, so I decided that the JSON impl
starts after a connection or session or whatever is established, and
just talks over the wire.