wip: Refactor `Transport` and add another impl
I realized that having the `Transport` API operate on raw bytes is not
very useful.  In order to implement this, the transport provider will at
the very least need to decode bytes to a String and then separate
requests by newlines, so it's more efficient to just make the transport
read/write String types directly.

I also implemented `Transport` for line-oriented streams wrapped in
`Framed`, which happens to be what I need in order to write some tests
that exercise the client and server.