wip: Big step forward in testing.
Router now has tests to exercise panic handling, all of which failed
until I implemented proper panic handling.

The integration test that exercises the handler that panics now also
works correctly since the router is catching the panic.

However, before I added panic handling, the integration test that caused
a panic was not failing as I expected, but hanging.  So now I need to do
more testing to simulate a panic in the event loop, and make sure it
gets handled properly so that at least clients don't hang.