wip: write test for bug in cancellation handling
I don't have a fix for this yet.  If the server shuts down while
handling a long-running method call, the client gets an error but the
error is from the client's own event loop which also shuts down when the
server connection drops; it doesn't actually get the response back from
the server about the server aborting, even though that response is
already locally in the client's pending incoming requests list.