wip: Nice refactor of event loop out into separate module
Also handle certain message types inline in the event loop instead of
spawning tasks.  Still spawn tasks for long-running operations.

Also factored out the spawning of the event loop future; it's now
returned along with the handle, and the caller is responsible for
polling it.  The `Client` and `Server` types do this automatically, but
I think over time something more clever will be needed, especially on
the server side for orderly shutdowns.