feat: Add compatibility tests for JSONRPC
This is probably way overkill for what's only going to be an internal
JSONRPC impl talking to other MCP impls, but it's my free time and I'll
waste it how I want.

This adds some nice compatibility tests where we drive the JSON RPC
service using JSON (and sometimes not JSON) strings defined in YAML
files.  In the process this revealed a few deviations from the spec or
just ugly things that needed cleaning up.