feat: Add request-level cancellation support to client and server requests
The MCP spec requires this, and it's best implemented at the JSON-RPC
level.

This touched a number of areas, and saw several new tests added.  Now
each request has a cancellation token, and the authors of handlers can,
if they go to extra effort, make their handlers smart enough to check
the cancellation token and cancel gracefully.  If handlers do not take
this extra effort, then the graceful shutdown timeout will apply after
which their futures will be aborted by the runtime.