refactor: Remove Arc and Mutex from ServiceConnection fields
In a surprise twist, the way that I designed the connection handler,
using an event loop, means that a single ServiceConnection does not need
to be `Sync`!  So I was doing this locking with Arc/Mutex unnecessarily.
Removing it cleans things up a bit and makes me happier with the event
loop implementation.