refactor: Refactor ServiceConnectionHandle and ServiceConnection to separate modules
The `service_connection.rs` module was over 3K lines and played havoc
with the LLM context window.  It was also getting hard to navigate
manually.  I'm refactoring it to separate out the
ServiceConnectionHandle and associated types and tests, from
ServiceConnection and its posse.  Both are going into submodules under
`service`.